# packages/shared/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/characters/{characterId}": {
      "get": {
        "operationId": "CharacterController_findOne",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific character by ID",
        "tags": [
          "characters"
        ]
      },
      "put": {
        "operationId": "CharacterController_update",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "description": "ID of the character to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCharacterRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Character updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a character",
        "tags": [
          "characters"
        ]
      },
      "delete": {
        "operationId": "CharacterController_delete",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a character",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/characters/{characterId}/inventory": {
      "post": {
        "operationId": "CharacterInventoryController_addItem",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item added to inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an item to character's inventory",
        "tags": [
          "character-inventory"
        ]
      }
    },
    "/api/characters/{characterId}/inventory/equip": {
      "post": {
        "operationId": "CharacterInventoryController_equipItem",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipInventoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Character updated with equipped item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Equip an item by definitionId (weapon only)",
        "tags": [
          "character-inventory"
        ]
      }
    },
    "/api/characters/{characterId}/inventory/{itemId}": {
      "patch": {
        "operationId": "CharacterInventoryController_updateItem",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventory item updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character or item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an item in character's inventory",
        "tags": [
          "character-inventory"
        ]
      },
      "delete": {
        "operationId": "CharacterInventoryController_removeItem",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveInventoryBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item removed from inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Character or item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an item from character's inventory",
        "tags": [
          "character-inventory"
        ]
      }
    },
    "/api/characters/{characterId}/inspiration/grant": {
      "post": {
        "operationId": "CharacterInspirationController_grant",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantInspirationBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inspiration granted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspirationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount"
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grant inspiration point(s) to a character",
        "tags": [
          "character-inspiration"
        ]
      }
    },
    "/api/characters/{characterId}/inspiration/spend": {
      "post": {
        "operationId": "CharacterInspirationController_spend",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Inspiration spent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspirationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No inspiration points available"
          },
          "404": {
            "description": "Character not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Spend an inspiration point",
        "tags": [
          "character-inspiration"
        ]
      }
    },
...
  },
  "info": {
    "title": "RPG Gen Portal",
    "description": "Prototype API for RPG portal using Gemini tokens",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TacticalStats": {
        "type": "object",
        "properties": {
          "vigor": {
            "type": "number",
            "description": "Vigor stat"
          },
          "finesse": {
            "type": "number",
            "description": "Finesse stat"
          },
          "mind": {
            "type": "number",
            "description": "Mind stat"
          },
          "survival": {
            "type": "number",
            "description": "Survival stat"
          }
        },
        "required": [
          "vigor",
          "finesse",
          "mind",
          "survival"
        ]
      },
...
      "RaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Race ID (unique identifier)",
            "example": "humain"
          },
          "name": {
            "type": "string",
            "description": "Display name",
            "example": "Humain"
          },
          "trait": {
            "type": "string",
            "description": "Trait name",
            "example": "Polyvalent"
          },
          "traitEffect": {
            "description": "Trait effect details",
            "allOf": [
              {
                "$ref": "#/components/schemas/TraitEffectResponseDto"
              }
            ]
          },
          "bonuses": {
            "description": "Stat bonuses",
            "allOf": [
              {
                "$ref": "#/components/schemas/RaceBonusesResponseDto"
              }
            ]
          },
          "descriptionForAi": {
            "type": "string",
            "description": "Description for AI",
            "example": "Polyvalent, gagne +1 PA au premier tour."
          },
          "color": {
            "type": "string",
            "description": "UI color (hex)",
            "example": "#3b82f6"
          },
          "icon": {
            "type": "string",
            "description": "Icon emoji",
            "example": "👤"
          }
        },
        "required": [
          "id",
          "name",
          "trait",
          "traitEffect",
          "bonuses",
          "color",
          "icon"
        ]
      },
      "SkillResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Skill name"
          },
          "proficient": {
            "type": "boolean",
            "description": "Is proficient in this skill"
          },
          "modifier": {
            "type": "number",
            "description": "Skill modifier"
          }
        }
      },
      "InventoryItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Item ID"
          },
          "definitionId": {
            "type": "string",
            "description": "Definition ID"
          },
          "name": {
            "type": "string",
            "description": "Item name"
          },
          "qty": {
            "type": "number",
            "description": "Quantity"
          },
          "description": {
            "type": "string",
            "description": "Item description"
          },
          "equipped": {
            "type": "boolean",
            "description": "Is equipped"
          },
          "meta": {
            "description": "Arbitrary item meta",
            "oneOf": [
              {
                "$ref": "#/components/schemas/WeaponMeta"
              },
              {
                "$ref": "#/components/schemas/ArmorMeta"
              },
              {
                "$ref": "#/components/schemas/ConsumableMeta"
              },
              {
                "$ref": "#/components/schemas/PackMeta"
              },
              {
                "$ref": "#/components/schemas/ToolMeta"
              }
            ]
          }
        },
        "required": [
          "definitionId",
          "name",
          "equipped",
          "meta"
        ]
      },
...
      "AptitudeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Aptitude ID (unique identifier)",
            "example": "frappe_simple"
          },
          "name": {
            "type": "string",
            "description": "Display name",
            "example": "Frappe simple"
          },
          "description": {
            "type": "string",
            "description": "Description",
            "example": "Une attaque basique"
          },
          "paCost": {
            "type": "number",
            "description": "Action points cost",
            "example": 2
          },
          "cooldown": {
            "type": "number",
            "description": "Cooldown in turns",
            "example": 0
          },
          "targeting": {
            "description": "Targeting information",
            "allOf": [
              {
                "$ref": "#/components/schemas/AptitudeTargetingResponseDto"
              }
            ]
          },
          "effects": {
            "description": "Effects list",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AptitudeEffectResponseDto"
            }
          },
          "icon": {
            "type": "string",
            "description": "Icon emoji",
            "example": "⚔️"
          }
        },
        "required": [
          "id",
          "name",
          "paCost",
          "targeting"
        ]
      },
...
      "CharacterResponseDto": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "description": "Unique character ID (UUID)"
          },
          "name": {
            "type": "string",
            "description": "Character name"
          },
          "physicalDescription": {
            "type": "string",
            "description": "Physical description of the character"
          },
          "race": {
            "description": "Character race (new system)",
            "allOf": [
              {
                "$ref": "#/components/schemas/RaceResponseDto"
              }
            ]
          },
          "hp": {
            "type": "number",
            "description": "Current hit points"
          },
          "hpMax": {
            "type": "number",
            "description": "Maximum hit points"
          },
          "totalXp": {
            "type": "number",
            "description": "Total experience points"
          },
          "skills": {
            "description": "Character skills",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResponseDto"
            }
          },
          "portrait": {
            "type": "string",
            "description": "Character portrait URL or base64"
          },
          "gender": {
            "type": "string",
            "description": "Character gender"
          },
          "inspirationPoints": {
            "type": "number",
            "description": "Inspiration points"
          },
          "isDeceased": {
            "type": "boolean",
            "description": "Whether character is deceased"
          },
          "diedAt": {
            "type": "string",
            "description": "Date of death (ISO string)"
          },
          "deathLocation": {
            "type": "string",
            "description": "Location where character died"
          },
          "state": {
            "type": "string",
            "description": "Character state",
            "enum": [
              "draft",
              "created"
            ]
          },
          "inventory": {
            "description": "Character inventory",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemDto"
            }
          },
          "aptitudes": {
            "description": "Character aptitudes (learned abilities)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AptitudeResponseDto"
            }
          },
          "className": {
            "type": "string",
            "description": "Character class (guerrier, rogue, mage)",
            "enum": [
              "guerrier",
              "rogue",
              "mage"
            ]
          },
          "level": {
            "type": "number",
            "description": "Character level (1-20)"
          },
          "raceId": {
            "type": "string",
            "description": "Race ID (humain, nain, elfe, dark_elfe, orc)",
            "enum": [
              "humain",
              "nain",
              "elfe",
              "dark_elfe",
              "orc"
            ]
          },
          "stats": {
            "description": "Tactical stats (vigor, finesse, mind, survival)",
            "allOf": [
              {
                "$ref": "#/components/schemas/TacticalStats"
              }
            ]
          },
          "pa": {
            "type": "number",
            "description": "Current action points"
          },
          "paMax": {
            "type": "number",
            "description": "Maximum action points"
          },
          "pm": {
            "type": "number",
            "description": "Current movement points"
          },
          "pmMax": {
            "type": "number",
            "description": "Maximum movement points"
          },
          "talentPoints": {
            "type": "number",
            "description": "Unspent talent points"
          },
          "voies": {
            "description": "Talent tree progression",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoieProgressDto"
            }
          }
        },
        "required": [
          "characterId",
          "isDeceased",
          "state"
        ]
      },
      "UpdateCharacterRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Character name"
          },
          "physicalDescription": {
            "type": "string",
            "description": "Physical description of the character"
          },
          "race": {
            "description": "Character race (new system)",
            "allOf": [
              {
                "$ref": "#/components/schemas/RaceResponseDto"
              }
            ]
          },
          "hp": {
            "type": "number",
            "description": "Current hit points"
          },
          "hpMax": {
            "type": "number",
            "description": "Maximum hit points"
          },
          "totalXp": {
            "type": "number",
            "description": "Total experience points"
          },
          "skills": {
            "description": "Character skills",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResponseDto"
            }
          },
          "portrait": {
            "type": "string",
            "description": "Character portrait URL or base64"
          },
          "gender": {
            "type": "string",
            "description": "Character gender"
          },
          "inspirationPoints": {
            "type": "number",
            "description": "Inspiration points"
          },
          "state": {
            "type": "string",
            "description": "Character state",
            "enum": [
              "draft",
              "created"
            ]
          },
          "inventory": {
            "description": "Character inventory",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemDto"
            }
          },
          "aptitudes": {
            "description": "Character aptitudes (learned abilities)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AptitudeResponseDto"
            }
          },
          "className": {
            "type": "string",
            "description": "Character class (guerrier, rogue, mage)",
            "enum": [
              "guerrier",
              "rogue",
              "mage"
            ]
          },
          "level": {
            "type": "number",
            "description": "Character level (1-20)"
          },
          "raceId": {
            "type": "string",
            "description": "Race ID (humain, nain, elfe, dark_elfe, orc)",
            "enum": [
              "humain",
              "nain",
              "elfe",
              "dark_elfe",
              "orc"
            ]
          },
          "stats": {
            "description": "Tactical stats (vigor, finesse, mind, survival)",
            "allOf": [
              {
                "$ref": "#/components/schemas/TacticalStats"
              }
            ]
          },
          "pa": {
            "type": "number",
            "description": "Current action points"
          },
          "paMax": {
            "type": "number",
            "description": "Maximum action points"
          },
          "pm": {
            "type": "number",
            "description": "Current movement points"
          },
          "pmMax": {
            "type": "number",
            "description": "Maximum movement points"
          },
          "talentPoints": {
            "type": "number",
            "description": "Unspent talent points"
          },
          "voies": {
            "description": "Talent tree progression",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoieProgressDto"
            }
          }
        }
      },
...
      "PackMeta": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Type discriminator for packs",
            "enum": [
              "pack"
            ]
          },
          "cost": {
            "type": "object",
            "description": "Item cost (gold pieces or formatted string)"
          },
          "weight": {
            "type": "object",
            "description": "Item weight in pounds"
          },
          "starter": {
            "type": "boolean",
            "description": "Whether this is a starter item"
          }
        }
      },
...
      "CreateInventoryItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Inventory item id (UUID). If provided, attempt to merge with existing item"
          },
          "definitionId": {
            "type": "string",
            "description": "Canonical definition id for this item"
          },
          "name": {
            "type": "string",
            "description": "Name for this inventory item"
          },
          "qty": {
            "type": "number",
            "description": "Quantity for this item"
          },
          "description": {
            "type": "string",
            "description": "Item description / notes"
          },
          "equipped": {
            "type": "boolean",
            "description": "If true this item is equipped"
          },
          "meta": {
            "description": "Arbitrary item meta",
            "oneOf": [
              {
                "$ref": "#/components/schemas/WeaponMeta"
              },
              {
                "$ref": "#/components/schemas/ArmorMeta"
              },
              {
                "$ref": "#/components/schemas/ConsumableMeta"
              },
              {
                "$ref": "#/components/schemas/PackMeta"
              },
              {
                "$ref": "#/components/schemas/ToolMeta"
              },
              {
                "$ref": "#/components/schemas/GenericMeta"
              }
            ]
          }
        },
        "required": [
          "definitionId"
        ]
      },
      "EquipInventoryDto": {
        "type": "object",
        "properties": {
          "definitionId": {
            "type": "string",
            "description": "Definition id of the item to equip",
            "example": "weapon-rapier"
          }
        },
        "required": [
          "definitionId"
        ]
      },
      "RemoveInventoryBodyDto": {
        "type": "object",
        "properties": {
          "qty": {
            "type": "number",
            "description": "Quantity to remove (-1 = remove all)"
          }
        }
      },
      "GrantInspirationBodyDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Amount of inspiration to grant (1-5)",
            "default": 1
          }
        }
      },
      "InspirationResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "inspirationPoints": {
            "type": "number",
            "description": "Updated inspiration points count"
          },
          "character": {
            "description": "Updated character",
            "allOf": [
              {
                "$ref": "#/components/schemas/CharacterResponseDto"
              }
            ]
          }
        },
        "required": [
          "ok",
          "inspirationPoints",
          "character"
        ]
      },
...
    }
  }
}